requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
markdown>=3.5.0
flask>=3.0.0
pandas>=2.0.0
duckdb>=0.9.0
openai>=1.0.0
gitpython>=3.1.0
python-dotenv>=1.0.0 